% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code3.R
\name{dag}
\alias{dag}
\title{Title}
\usage{
dag(ct = NULL, data = NULL, num_p, type = c("discrete", "gaussian",
  "zero_inflated"), clustersize)
}
\arguments{
\item{ct}{the contingincy table converted from data}

\item{data}{dataset}

\item{num_p}{the maximum number of parents}

\item{type}{a character representing the type of your data. Choosing from "discrete","gausssian","zero_inflated"}

\item{clustersize}{the number of parallel}

\item{DAG}{represents the best bayesian network}

\item{score}{represents the score of the bayesian network}
}
\description{
This function is used to find the best bayesian network for dicrete, gaussian and zero-inflated data.
}
\details{
This function is used to find the best bayesian network for dicrete, gaussian and zero-inflated data by using dynamic programming.
}
\author{
Yejiong Zhu
}
